+++ examples/button.rs
@@ -2,20 +2,24 @@ use std::process::ExitCode;
 
 use termal::eprintcln;
 use termint::{
+    Error,
     enums::{BorderType, Color},
     geometry::Constraint,
-    style::Style,
+    prelude::TextAlign,
+    style::{Style, Stylize},
     term::{
-        backend::{Event, KeyCode, KeyEvent, MouseButton},
         Action, Application, Frame, Term,
+        backend::{Event, KeyCode, KeyEvent, MouseButton},
     },
     widgets::{Block, Button, Element, Layout, Spacer, ToSpan},
-    Error,
 };
 
 const BG: Color = Color::Hex(0x02081e);
 const BORDER: Color = Color::Hex(0x535C91);
 const FG: Color = Color::Hex(0xc3c1f4);
+const INC: Color = Color::Hex(0xA3DC9A);
+const DEC: Color = Color::Hex(0xEA7B7B);
+const BTN: Color = Color::Hex(0x98A1BC);
 
 fn main() -> ExitCode {
     if let Err(e) = run() {
@@ -32,39 +36,55 @@ fn run() -> Result<(), Error> {
 
 #[derive(Clone)]
 enum Message {
+    Increment,
+    Decrement,
+    Reset,
     Quit,
 }
 
 #[derive(Default)]
-struct App;
+struct App {
+    count: isize,
+}
 
 impl Application for App {
     type Message = Message;
 
     fn view(&self, _frame: &Frame) -> Element<Self::Message> {
         let mut block = Block::vertical()
-            .title("Centering")
+            .title("Button Example")
             .border_type(BorderType::Thicker)
             .border_style(Style::new().bg(BG).fg(BORDER))
             .style(Style::new().bg(BG).fg(FG));
 
-        let left = Button::new("Left Click")
-            .style(Style::new().bg(Color::Cyan).fg(BG))
+        let counter = format!("Counter: {}", self.count).bold();
+
+        let inc_btn = Button::new("+".align(TextAlign::Center))
+            .style((BG, INC))
             .padding((1, 2))
-            .on_click(Message::Quit);
-        let right = Button::new("Right Click")
-            .style(Style::new().bg(Color::Cyan).fg(BG))
+            .on_click(Message::Increment);
+        let dec_btn = Button::new("-".align(TextAlign::Center))
+            .style((BG, DEC))
             .padding((1, 2))
+            .on_click(Message::Decrement);
+
+        let btn = Button::new("Left = Reset, Right = Quit")
+            .style((BG, BTN))
+            .padding((1, 2))
+            .on_click(Message::Reset)
             .on_press(MouseButton::Right, Message::Quit);
-        let mut buttons = Layout::horizontal();
-        buttons.push(left, 14);
-        buttons.push(Spacer::new(), 2);
-        buttons.push(right, 15);
+
+        let mut buttons = Layout::horizontal().center();
+        buttons.push(inc_btn, 11);
+        buttons.push(Spacer::new(), 8);
+        buttons.push(dec_btn, 11);
 
         let mut wrapper = Layout::vertical();
-        wrapper.push("Press a button to quit...", 0..);
+        wrapper.push(counter, 0..);
         wrapper.push(Spacer::new(), 1);
         wrapper.push(buttons, 0..);
+        wrapper.push(Spacer::new(), 1);
+        wrapper.push(btn, 0..);
 
         let mut center = Layout::horizontal().center();
         center.push(wrapper, 0..);
@@ -86,8 +106,12 @@ impl Application for App {
 
     fn message(&mut self, message: Self::Message) -> Action {
         match message {
-            Message::Quit => Action::QUIT,
+            Message::Increment => self.count += 1,
+            Message::Decrement => self.count -= 1,
+            Message::Reset => self.count = 0,
+            Message::Quit => return Action::QUIT,
         }
+        Action::RENDER
     }
 }